+++ scripts/check-package-lock.ts
@@ -1,0 +1,3 @@
+import { PackageLockValidator } from './components/package-lock-validator.js';
+
+(async () => await new PackageLockValidator('package-lock.json').validate())();